First successful compilation of both ios and mac lib. Force define TARGET_OS_IPHONE and TARGET_OS_MAC in prefix headers of each target

+++ CLToolkit/Networking/RACHTTPClient.m
@@ -32,7 +32,7 @@ static void LogHTTPResponse(int tag, NSHTTPURLResponse *response) {
     NSString *path = response.URL.path;
     id body = [response json];
     NSString *tagStr = [$str(@"<-  R%d     ", tag) sliceTill:8];
-    LogDebug(@"http", @"%@ %ld %@\n Body: %@\n Headers: %@", tagStr, response.statusCode, path, body, headers);
+    LogDebug(@"http", @"%@ %ld %@\n Body: %@\n Headers: %@", tagStr, (long)response.statusCode, path, body, headers);
     LoggerFlush(NULL, NO);
 }
 static void LogHTTPError(int tag, NSError *error) {
